refactoring and simplifying controller structure
