+++ model/src/main/scala/uk/gov/ons/addressIndex/model/config/AddressIndexConfig.scala
@@ -18,8 +18,6 @@ object ElasticSearchConfig {
     cluster = "ons-cluster",
     local = false,
     indexes = IndexesConfig(
-      pafIndex = "paf/address",
-      nagIndex = "nag/address",
       hybridIndex = "hybrid/address"
     ),
     shield = ShieldConfig(
@@ -51,8 +49,6 @@ object AddressIndexConfig {
 }
 
 case class IndexesConfig(
-  pafIndex: String,
-  nagIndex: String,
   hybridIndex: String
 )
