+++ server/conf/application.conf
@@ -27,10 +27,6 @@ addressIndex {
       ssl = ${?ONS_AI_API_SHIELD_SSL}
     }
     indexes {
-      pafIndex = "paf/address"
-      pafIndex = ${?ONS_AI_API_PAF_INDEX}
-      nagIndex = "nag/address"
-      nagIndex = ${?ONS_AI_API_NAG_INDEX}
       hybridIndex = "hybrid/address"
       hybridIndex = ${?ONS_AI_API_HYBRID_INDEX}
     }
